moved keep alive logic back into keep_alive_system
