+++ src/bin/src/systems/ticking_system.rs
@@ -11,7 +11,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::time::Instant;
-use tracing::{debug, error, info};
+use tracing::{debug, info};
 pub struct TickingSystem;
 
 static KILLED: AtomicBool = AtomicBool::new(false);
@@ -25,63 +25,6 @@ impl System for TickingSystem {
             let required_end = Instant::now() + Duration::from_millis(50);
             // TODO handle error
             let res = TickEvent::trigger(TickEvent::new(tick), state.clone()).await;
-
-            let current_time = std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .expect("Time went backwards")
-                .as_millis() as i64;
-
-            let fifteen_seconds_ms = 15000; // 15 seconds in milliseconds
-
-            let query = state
-                .universe
-                .query::<(&mut StreamWriter, &ConnectionState, &KeepAlive)>()
-                .into_entities()
-                .into_iter()
-                .filter_map(|entity| {
-                    let conn_state = state.universe.get::<ConnectionState>(entity).ok()?;
-                    let keep_alive = state.universe.get_mut::<KeepAlive>(entity).ok()?;
-
-                    if matches!(*conn_state, ConnectionState::Play)
-                        && (current_time - keep_alive.id) > fifteen_seconds_ms
-                    {
-                        Some(entity)
-                    } else {
-                        None
-                    }
-                })
-                .collect::<Vec<_>>();
-            if query.len() > 0 {
-                info!("there are {:?} players to keep alive", query.len());
-            }
-
-            let packet = KeepAlivePacket::default();
-            let packet = {
-                let mut buffer = Vec::new();
-                if packet
-                    .encode(&mut buffer, &NetEncodeOpts::WithLength)
-                    .is_err()
-                {
-                    error!("Error encoding keep alive packet");
-                }
-                buffer
-            };
-            tokio::spawn(futures::stream::iter(query.into_iter()).fold(
-                (state.clone(), packet),
-                move |(state, packet), entity| async move {
-                    if let Ok(mut writer) = state.universe.get_mut::<StreamWriter>(entity) {
-                        if let Err(e) = writer
-                            .send_packet(&packet.as_slice(), &NetEncodeOpts::None)
-                            .await
-                        {
-                            error!("Error sending update_time packet: {}", e);
-                        }
-                    }
-                    debug!("Sent keep alive packet to {}", entity);
-                    (state, packet)
-                },
-            ));
-
             if res.is_err() {
                 debug!("error handling tick event: {:?}", res);
             }
